+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## Bugfixes
 - Fixed removing enchantments disallowing items to be enchanted. [#5](https://github.com/GreenhouseModding/enchantment-disable-tag/pull/5) - theendercore
 - Fixed removed enchantments appearing in the enchanting table. [#5](https://github.com/GreenhouseModding/enchantment-disable-tag/pull/5) - theendercore
+- Fixed slots without any enchantments showing up in the enchanting table.